Restructure board code
  ref #65
